backend/requirements.txt
alembic==1.16.5
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cryptography==46.0.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyotp==2.9.0
pyparsing==3.2.5
python-dotenv==1.1.1
python-http-client==3.3.7
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
SQLAlchemy==2.0.42
tomli==2.2.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
